Create Documentation.md
according to the last HTML documentation, the entity is deprecated and only added in IFC4x3_RC1. As a general topic, to be discussed in a general issue, I would argue, that any entity, that never has been part of an officially released IFC version, should not remain in the schema, but get deleted.