Update Documentation.md
added the note, that the Axis1 is not identical to x-axis for the swept solid at this point, x-axis is the orthogonal complement.

there seems to be a markdown generation area, all "*" or (star) characters get excaped with a "\" and thereby destroying the expected markdown style (a bullet list or a bold character set). As it happens everywhere, it looks to be a systematic error.